+++ packages/color/src/formatter.js
@@ -1,4 +1,4 @@
-import { clone } from "./utils/index";
+import { clone, round } from "./utils/index";
 
 /**
  * @method format
@@ -31,6 +31,10 @@ export function format(obj, type, defaultColor = "rgba(0, 0, 0, 0)") {
     return rgb(obj, defaultColor);
   } else if (type == "hsl") {
     return hsl(obj);
+  } else if (type == "hsv") {
+    return hsv(obj);
+  } else if (type == "cmyk") {
+    return cmyk(obj);
   }
 
   return `${type}(${obj.r},${obj.g},${obj.b})`;
@@ -92,12 +96,42 @@ export function rgb(obj, defaultColor = "rgba(0, 0, 0, 0)") {
 
 export function hsl(obj) {
   if (Array.isArray(obj)) {
-    obj = { r: obj[0], g: obj[1], b: obj[2], a: obj[3] };
+    obj = { h: obj[0], s: obj[1], l: obj[2], a: obj[3] };
   }
 
   if (obj.a == 1 || typeof obj.a === "undefined") {
-    return `hsl(${obj.h},${obj.s}%,${obj.l}%)`;
+    return `hsl(${obj.h}, ${obj.s}%, ${obj.l}%)`;
   } else {
-    return `hsla(${obj.h},${obj.s}%,${obj.l}%,${obj.a})`;
+    return `hsla(${obj.h}, ${obj.s}%, ${obj.l}%, ${obj.a})`;
   }
+}
+
+export function hsv(obj) {
+  if (Array.isArray(obj)) {
+    obj = { h: obj[0], s: obj[1], v: obj[2], a: obj[3] };
+  }
+
+  const h = round(obj.h);
+  const a = obj.a;
+  const s = round(obj.s * 100);
+  const v = round(obj.v * 100);
+
+  if (obj.a == 1 || typeof obj.a === "undefined") {
+    return `hsv(${h}, ${s}%, ${v}%)`;
+  } else {
+    return `hsva(${h}, ${s}%, ${v}%, ${a})`;
+  }
+}
+
+export function cmyk(obj) {
+  if (Array.isArray(obj)) {
+    obj = { c: obj[0], m: obj[1], y: obj[2], k: obj[3] };
+  }
+
+  const c = round(obj.c * 100, 100);
+  const m = round(obj.m * 100, 100);
+  const y = round(obj.y * 100, 100);
+  const k = round(obj.k * 100, 100);
+
+  return `cmyk(${c}%,${m}%,${y}%,${k}%)`;
 }