+++ apps/ui-dev/CHANGELOG.md
@@ -1,5 +1,11 @@
 # ui-dev
 
+## 1.0.13
+
+### Patch Changes
+
+- update color/typography for design documentation
+
 ## 1.0.12
 
 ### Patch Changes